fix(cli/migrate): do not load & print env file for all commands
Required for `migrate diff`.

2 new functions:
- getSchemaPathAndPrint
- loadEnvFileAndPrint

With refactor of some duplicated code
