+++ packages/cli/src/Doctor.ts
@@ -1,7 +1,6 @@
 import type { Command } from '@prisma/sdk'
 import {
   arg,
-  getSchemaPath,
   getDMMF,
   getConfig,
   IntrospectionEngine,
@@ -10,6 +9,7 @@ import {
   format,
   HelpError,
   canConnectToDatabase,
+  loadEnvFileAndPrint,
 } from '@prisma/sdk'
 import chalk from 'chalk'
 import fs from 'fs'
@@ -17,6 +17,7 @@ import path from 'path'
 import { promisify } from 'util'
 import type { DMMF } from '@prisma/generator-helper'
 import equal from 'fast-deep-equal'
+import { getSchemaPathAndPrint } from '@prisma/migrate'
 
 const readFile = promisify(fs.readFile)
 type IncorrectFieldTypes = Array<{
@@ -67,21 +68,9 @@ ${chalk.bold('Examples')}
       return this.help()
     }
 
-    const schemaPath = await getSchemaPath(args['--schema'])
+    loadEnvFileAndPrint(args['--schema'])
 
-    if (!schemaPath) {
-      throw new Error(
-        `Could not find a ${chalk.bold(
-          'schema.prisma',
-        )} file that is required for this command.\nYou can either provide it with ${chalk.greenBright(
-          '--schema',
-        )}, set it as \`prisma.schema\` in your package.json or put it into the default location ${chalk.greenBright(
-          './prisma/schema.prisma',
-        )} https://pris.ly/d/prisma-schema-location`,
-      )
-    }
-
-    console.log(chalk.dim(`Prisma schema loaded from ${path.relative(process.cwd(), schemaPath)}`))
+    const schemaPath = await getSchemaPathAndPrint(args['--schema'])
 
     const schema = await readFile(schemaPath, 'utf-8')
     const localDmmf = await getDMMF({ datamodel: schema })
