+++ packages/cli/src/Format.ts
@@ -1,9 +1,9 @@
+import { getSchemaPathAndPrint } from '@prisma/migrate'
 import type { Command } from '@prisma/sdk'
-import { arg, format, formatms, formatSchema, getDMMF, getSchemaPath, HelpError } from '@prisma/sdk'
+import { arg, format, formatms, formatSchema, getDMMF, HelpError } from '@prisma/sdk'
 import chalk from 'chalk'
 import fs from 'fs'
 import os from 'os'
-import path from 'path'
 
 /**
  * $ prisma format
@@ -52,21 +52,7 @@ Or specify a Prisma schema path
       return this.help()
     }
 
-    const schemaPath = await getSchemaPath(args['--schema'])
-
-    if (!schemaPath) {
-      throw new Error(
-        `Could not find a ${chalk.bold(
-          'schema.prisma',
-        )} file that is required for this command.\nYou can either provide it with ${chalk.greenBright(
-          '--schema',
-        )}, set it as \`prisma.schema\` in your package.json or put it into the default location ${chalk.greenBright(
-          './prisma/schema.prisma',
-        )} https://pris.ly/d/prisma-schema-location`,
-      )
-    }
-
-    console.log(chalk.dim(`Prisma schema loaded from ${path.relative(process.cwd(), schemaPath)}`))
+    const schemaPath = await getSchemaPathAndPrint(args['--schema'])
 
     let output = await formatSchema({
       schemaPath,
