+++ packages/cli/src/Validate.ts
@@ -1,8 +1,9 @@
 import type { Command } from '@prisma/sdk'
-import { arg, format, getConfig, getDMMF, getSchemaPath, HelpError } from '@prisma/sdk'
+import { loadEnvFileAndPrint } from '@prisma/sdk'
+import { arg, format, getConfig, getDMMF, HelpError } from '@prisma/sdk'
 import chalk from 'chalk'
 import fs from 'fs'
-import path from 'path'
+import { getSchemaPathAndPrint } from '@prisma/migrate'
 
 /**
  * $ prisma validate
@@ -50,21 +51,9 @@ ${chalk.bold('Examples')}
       return this.help()
     }
 
-    const schemaPath = await getSchemaPath(args['--schema'])
+    loadEnvFileAndPrint(args['--schema'])
 
-    if (!schemaPath) {
-      throw new Error(
-        `Could not find a ${chalk.bold(
-          'schema.prisma',
-        )} file that is required for this command.\nYou can either provide it with ${chalk.greenBright(
-          '--schema',
-        )}, set it as \`prisma.schema\` in your package.json or put it into the default location ${chalk.greenBright(
-          './prisma/schema.prisma',
-        )} https://pris.ly/d/prisma-schema-location`,
-      )
-    }
-
-    console.log(chalk.dim(`Prisma schema loaded from ${path.relative(process.cwd(), schemaPath)}`))
+    const schemaPath = await getSchemaPathAndPrint(args['--schema'])
 
     const schema = fs.readFileSync(schemaPath, 'utf-8')
