+++ packages/cli/src/Version.ts
@@ -1,6 +1,7 @@
 import { getCliQueryEngineBinaryType } from '@prisma/engines'
 import { getPlatform } from '@prisma/get-platform'
 import type { Command } from '@prisma/sdk'
+import { loadEnvFileAndPrint } from '@prisma/sdk'
 import {
   arg,
   BinaryType,
@@ -65,6 +66,8 @@ export class Version implements Command {
     if (args['--help']) {
       return this.help()
     }
+
+    loadEnvFileAndPrint()
 
     const platform = await getPlatform()
     const cliQueryEngineBinaryType = getCliQueryEngineBinaryType()
