+++ packages/cli/src/bin.ts
@@ -3,16 +3,7 @@
 // hides ExperimentalWarning: The fs.promises API is experimental
 process.env.NODE_NO_WARNINGS = '1'
 
-import {
-  arg,
-  getCLIPathHash,
-  getProjectHash,
-  getSchema,
-  getConfig,
-  tryLoadEnvs,
-  getEnvPaths,
-  parseEnvValue,
-} from '@prisma/sdk'
+import { arg, getCLIPathHash, getProjectHash, getSchema, getConfig, parseEnvValue } from '@prisma/sdk'
 import chalk from 'chalk'
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires, @typescript-eslint/no-unsafe-assignment
@@ -53,20 +44,6 @@ const args = arg(
   true,
 )
 
-//
-// Read .env file only if next to schema.prisma
-//
-// if the CLI is called without any command like `prisma` we can ignore .env loading
-if (commandArray.length) {
-  try {
-    const envPaths = getEnvPaths(args['--schema'])
-    const envData = tryLoadEnvs(envPaths, { conflictCheck: 'error' })
-    envData && envData.message && console.log(envData.message)
-  } catch (e) {
-    handleIndividualError(e)
-  }
-}
-
 /**
  * Dependencies
  */
@@ -98,6 +75,15 @@ import { isCurrentBinInstalledGlobally } from '@prisma/sdk'
 import { Validate } from './Validate'
 import { Format } from './Format'
 import { Doctor } from './Doctor'
+/*
+  When running bin.ts with ts-node with DEBUG="*"
+  This error shows and blocks the execution
+  Quick hack is to comment the Studio import and usage to use the CLI without building it...
+
+  prisma:cli Error: Cannot find module '@prisma/sdk'
+  prisma:cli Require stack:
+  prisma:cli - /Users/j42/Dev/prisma-meow/node_modules/.pnpm/@prisma+studio-pcw@0.456.0/node_modules/@prisma/studio-pcw/dist/index.js
+*/
 import { Studio } from './Studio'
 import { Telemetry } from './Telemetry'
 import { printUpdateMessage } from './utils/printUpdateMessage'
@@ -142,7 +128,6 @@ async function main(): Promise<number> {
        * @deprecated since version 2.30.0, use `db pull` instead (renamed)
        */
       introspect: DbPull.new(),
-      dev: Dev.new(),
       studio: Studio.new(),
       generate: Generate.new(),
       version: Version.new(),
@@ -150,6 +135,8 @@ async function main(): Promise<number> {
       format: Format.new(),
       doctor: Doctor.new(),
       telemetry: Telemetry.new(),
+      // TODO remove Legacy
+      dev: Dev.new(),
     },
     [
       'version',
@@ -157,7 +144,6 @@ async function main(): Promise<number> {
       'migrate',
       'db',
       'introspect',
-      'dev',
       'studio',
       'generate',
       'validate',
