+++ packages/migrate/src/commands/MigrateReset.ts
@@ -1,16 +1,17 @@
 import type { Command } from '@prisma/sdk'
-import { arg, format, getSchemaPath, HelpError, isError, isCi, logger } from '@prisma/sdk'
+import { loadEnvFileAndPrint } from '@prisma/sdk'
+import { arg, format, getSchemaPath, HelpError, isError, isCi } from '@prisma/sdk'
 import chalk from 'chalk'
-import path from 'path'
 import prompt from 'prompts'
 import { Migrate } from '../Migrate'
 import { ExperimentalFlagWithMigrateError, EarlyAccessFeatureFlagWithMigrateError } from '../utils/flagErrors'
-import { NoSchemaFoundError, MigrateResetEnvNonInteractiveError } from '../utils/errors'
+import { MigrateResetEnvNonInteractiveError } from '../utils/errors'
 import { printFilesFromMigrationIds } from '../utils/printFiles'
 import { throwUpgradeErrorIfOldMigrate } from '../utils/detectOldMigrate'
 import { ensureDatabaseExists } from '../utils/ensureDatabaseExists'
 import { printDatasource } from '../utils/printDatasource'
 import { executeSeedCommand, verifySeedConfigAndReturnMessage, getSeedCommandFromPackageJson } from '../utils/seed'
+import { getSchemaPathAndPrint } from '../utils/getSchemaPathAndPrint'
 
 export class MigrateReset implements Command {
   public static new(): MigrateReset {
@@ -74,13 +75,9 @@ ${chalk.bold('Examples')}
       throw new EarlyAccessFeatureFlagWithMigrateError()
     }
 
-    const schemaPath = await getSchemaPath(args['--schema'])
+    loadEnvFileAndPrint(args['--schema'])
 
-    if (!schemaPath) {
-      throw new NoSchemaFoundError()
-    }
-
-    console.info(chalk.dim(`Prisma schema loaded from ${path.relative(process.cwd(), schemaPath)}`))
+    const schemaPath = await getSchemaPathAndPrint(args['--schema'])
 
     await printDatasource(schemaPath)
