+++ packages/sdk/src/index.ts
@@ -70,3 +70,4 @@ export { getPlatform, getNodeAPIName } from '@prisma/get-platform'
 export type { Platform } from '@prisma/get-platform'
 export { platformRegex } from './utils/platformRegex'
 export { jestConsoleContext, jestContext } from './utils/jestContext'
+export { loadEnvFileAndPrint } from './utils/loadEnvFileAndPrint'
